+++ realm/realm-library/src/main/java/io/realm/Realm.java
@@ -40,6 +40,7 @@ import java.util.Scanner;
 import java.util.Set;
 import java.util.concurrent.Future;
 
+import io.realm.annotations.internal.OptionalAPI;
 import io.realm.exceptions.RealmException;
 import io.realm.exceptions.RealmInvalidDatabaseException;
 import io.realm.exceptions.RealmIOException;
@@ -141,6 +142,7 @@ public final class Realm extends BaseRealm {
      * {@inheritDoc}
      */
     @Override
+    @OptionalAPI(dependencies = {"rx.Observable"})
     public Observable<Realm> asObservable() {
         return configuration.getRxFactory().from(this);
     }