+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 * A `RealmInvalidDatabaseException` is thrown instead of `IllegalStateException` when opening a corrupted Realm file, a non-Realm file, a Realm file created by newer version of Realm, or an encrypted Realm file with a wrong key.
 
+### Enhancements
+
+* Enhanced `Table.toString()` to show a PrimaryKey field details (#2903).
+
 ## 1.0.1
 
 ### Bug fixes
@@ -13,7 +17,7 @@
 
 ### Enhancements
 
-* Enhanced `Table.toString()` to show a PrimaryKey field details (#2903).
+* Removes RxJava related APIs during bytecode transforming to make RealmObject plays well with reflection when rx.Observable doesn't exist.
 
 ## 1.0.0